refactor: ContentWrap 사용하는 컴포넌트 import 경로 수정
